Fleshed out infix ops proposal
The idea is that the conventional comparison and arithmetic operands are used but whenever a Rational is at a given side of the operator, then the operator is 'protected' by a '|' character.  This both avoids clashes with Basics operators and also gives a method of predicting what the operator is with either one or two Rationals in a given context.